# docs/fibersimple/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This about user location details",
        "title": "Fiber Swagger API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/v1/nearest/user": {
            "post": {
                "description": "it give nearest 10 user.",
                "consumes": [
                    "application/octet-stream"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "root"
                ],
                "summary": "get Get Nearest User.",
                "parameters": [
                    {
                        "description": "user cooridnates",
                        "name": "coordinates",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Coordinates"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.UserCoordinateItem"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.Coordinates": {
            "type": "object",
            "properties": {
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
        "types.UserCoordinateItem": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "number"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
    }
}